refactor(backend): extract shortform route helpers
